Moved examples in a tidy folder
