+++ website/docs/migration/from_state_notifier.mdx
@@ -2,13 +2,13 @@
 title: From `StateNotifier` and `ChangeNotifier`
 ---
 
-import buildIinit from "./build_init";
-import familyAndDispose from "./family_and_dispose";
-import asyncNotifier from "./async_notifier";
-import addListener from "./add_listener";
-import fromChangeNotifier from "./from_change_notifier";
-import fromStateProvider from "./from_state_provider";
-import oldLifecycles from "./old_lifecycles";
+import buildIinit from "./from_state_notifier/build_init";
+import familyAndDispose from "./from_state_notifier/family_and_dispose";
+import asyncNotifier from "./from_state_notifier/async_notifier";
+import addListener from "./from_state_notifier/add_listener";
+import fromChangeNotifier from "./from_state_notifier/from_change_notifier";
+import fromStateProvider from "./from_state_notifier/from_state_provider";
+import oldLifecycles from "./from_state_notifier/old_lifecycles";
 import { AutoSnippet } from "../../src/components/CodeSnippet";
 
 Along with [Riverpod 2.0](https://pub.dev/packages/flutter_riverpod/changelog#200), new classes
