+++ website/docs/migration/from_state_notifier/build_init/build_init.dart
@@ -4,7 +4,7 @@ import 'dart:math';
 
 import 'package:riverpod_annotation/riverpod_annotation.dart';
 
-import '../utils.dart';
+import '../../utils.dart';
 
 part 'build_init.g.dart';
