+++ website/docs/migration/from_state_notifier/family_and_dispose/family_and_dispose.dart
@@ -4,7 +4,7 @@ import 'dart:math';
 
 import 'package:riverpod_annotation/riverpod_annotation.dart';
 
-import '../utils.dart';
+import '../../utils.dart';
 
 part 'family_and_dispose.g.dart';
